Add River Meters to Blue Line Key

Adds distances (rm) and spatial coordinates (geometry) of regularly spaced points along blue line key (blk). All distances which are in meters are from the river mouth.

fwa_add_rms_to_blk(
  x,
  interval = 1000,
  start = 0,
  end = Inf,
  epsg = getOption("fwa.epsg", 3005)
)

Arguments

x
A data frame with integer column blk.

interval
A whole numeric of the distance between points.

start
A whole numeric of the start distance.

end
An integer of the end distance.

epsg
A positive whole number of EPSG projection for the coordinates.

Value

An sf tibble with the columns of x plus integer column rm and sf column geometry.

See also

Examples

fwa_add_rms_to_blk(data.frame(blk = 356308001))
#> Simple feature collection with 20 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 1290725 ymin: 464161.2 xmax: 1301656 ymax: 469483.5
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 20 × 4
#>          blk    rm elevation           geometry
#>        <dbl> <int>     <dbl>        <POINT [m]>
#>  1 356308001     0      9.63 (1290725 469483.5)
#>  2 356308001  1000     10.1  (1291542 469173.9)
#>  3 356308001  2000      7.5  (1292439 468922.9)
#>  4 356308001  3000      9.69 (1292909 468665.1)
#>  5 356308001  4000      9.78 (1293268 467977.4)
#>  6 356308001  5000      9.78 (1293158 467231.2)
#>  7 356308001  6000      9.78 (1293794 467668.3)
#>  8 356308001  7000      9.78 (1294079 467938.4)
#>  9 356308001  8000     10    (1294497 468239.5)
#> 10 356308001  9000     10.5  (1294708 468081.9)
#> 11 356308001 10000     10.5  (1295576 467912.1)
#> 12 356308001 11000     10.5  (1296011 467111.8)
#> 13 356308001 12000     10.5  (1296414 466358.5)
#> 14 356308001 13000     10.5  (1297305 465976.3)
#> 15 356308001 14000     10.5  (1298145 465435.2)
#> 16 356308001 15000     10.5  (1298990 464903.5)
#> 17 356308001 16000     11.5  (1299886 465109.3)
#> 18 356308001 17000     48.0  (1300649 465508.7)
#> 19 356308001 18000    161.   (1301136 464758.4)
#> 20 356308001 19000    210.   (1301656 464161.2)