FWA get named_streams

Get named stream features within watershed.

fwa_get_named_streams(
  watershed,
  intersect = FALSE,
  epsg = getOption("fwa.epsg", 3005),
  limit = getOption("fwa.limit", 10000)
)

Arguments

watershed
A single row sf object or sfc object of the watershed polygon (sfc geometry must inherit from sfc_POLYGON).

intersect
A flag specifying whether to intersect the individual features with the watershed. Useful if the watershed is not a true watershed.

epsg
A positive whole number of the epsg to return the collection to.

limit
A positive whole number indicating the maximum number of features to return.

Value

A sf object

See also

Examples

if (FALSE) {
watershed <- fwa_blue_line_key_to_watershed(356308001)
fwa_get_named_streams(watershed)
}