FWA rkm

Provided a blue_line_key and starting distance upstream from river mouth, return Rkm points along an interval.

fwa_rkm(
  blue_line_key,
  interval = 1000,
  distance_upstream = 0,
  epsg = 4326,
  limit = 500
)

Arguments

blue_line_key
An integer of the stream blue line key.

interval
An integer of the interval distance in meters.

distance_upstream
An integer of the distance in meters upstream from the river mouth to start from.

epsg
A positive whole number of the epsg to transform features to.

limit
A positive whole number of the maximum number of features to return.

Value

A sf object

Examples

fwa_rkm(356308001)

#> Simple feature collection with 20 features and 2 fields
#> geometry type:  POINT
#> dimension:      XYZ
#> bbox:           xmin: -122.0155 ymin: 49.11457 xmax: -121.8703 ymax: 49.16804
#> z_range:        zmin: 7.5 zmax: 209.8951
#> geographic CRS: WGS 84
#> First 10 features:
#>    blue_line_key rkm                       geometry
#> 1      356308001   0 POINT Z (-122.0155 49.16804...
#> 2      356308001   1 POINT Z (-122.0045 49.16484...
#> 3      356308001   2 POINT Z (-121.9925 49.16213...
#> 4      356308001   3 POINT Z (-121.9862 49.15957...
#> 5      356308001   4 POINT Z (-121.9819 49.15321...
#> 6      356308001   5 POINT Z (-121.984 49.14656 ...
#> 7      356308001   6 POINT Z (-121.9749 49.15016...
#> 8      356308001   7 POINT Z (-121.9708 49.15244...
#> 9      356308001   8 POINT Z (-121.9649 49.15493...
#> 10     356308001   9 POINT Z (-121.9621 49.15341...