Get River Metre

Provided a blue_line_key and start and possibly end distance upstream from river mouth, return river metre points at regular intervals.

fwa_rm(
  blue_line_key,
  interval = 1000,
  start = 0,
  end = NULL,
  epsg = getOption("fwa.epsg", 3005),
  limit = getOption("fwa.limit", 10000)
)

Arguments

blue_line_key
An integer of the stream blue line key.

interval
An integer of the interval distance in meters.

start
An integer of the distance in meters upstream from the river mouth to start from.

end
An integer of the distance in meters upstream from the river mouth to start from.

epsg
A positive whole number of the epsg to transform features to.

limit
A positive whole number of the maximum number of features to return.

Value

A sf object

See also

Examples

fwa_rm(356308001)
#> Simple feature collection with 20 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XYZ
#> Bounding box:  xmin: 1290725 ymin: 464161.2 xmax: 1301656 ymax: 469483.5
#> z_range:       zmin: 7.5 zmax: 209.8951
#> Projected CRS: NAD83 / BC Albers
#> First 10 features:
#>    blue_line_key   rm                       geometry
#> 1      356308001    0 POINT Z (1290725 469483.5 9...
#> 2      356308001 1000 POINT Z (1291542 469173.9 1...
#> 3      356308001 2000 POINT Z (1292439 468922.9 7.5)
#> 4      356308001 3000 POINT Z (1292909 468665.1 9...
#> 5      356308001 4000 POINT Z (1293268 467977.4 9...
#> 6      356308001 5000 POINT Z (1293158 467231.2 9...
#> 7      356308001 6000 POINT Z (1293794 467668.3 9...
#> 8      356308001 7000 POINT Z (1294079 467938.4 9...
#> 9      356308001 8000  POINT Z (1294497 468239.5 10)
#> 10     356308001 9000 POINT Z (1294708 468081.9 1...