Get Rkm

Provided a blue_line_key and start and possibly end distance upstream from river mouth, return river kilometre points at regular intervals.

fwa_rkm(
  blue_line_key,
  interval = 1000,
  start = 0,
  end = NULL,
  epsg = getOption("fwa.epsg", 3005),
  limit = getOption("fwa.limit", 10000)
)

Arguments

blue_line_key
An integer of the stream blue line key.

interval
An integer of the interval distance in meters.

start
An integer of the distance in meters upstream from the river mouth to start from.

end
An integer of the distance in meters upstream from the river mouth to start from.

epsg
A positive whole number of the epsg to transform features to.

limit
A positive whole number of the maximum number of features to return.

Value

A sf object

See also

Examples

fwa_rkm(356308001)
#> Warning: `fwa_rkm()` was deprecated in fwatlasbc 0.0.0.9004.
#> Please use `fwa_rm()` instead.
#> Simple feature collection with 20 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XYZ
#> Bounding box:  xmin: 1290725 ymin: 464161.2 xmax: 1301656 ymax: 469483.5
#> z_range:       zmin: 7.5 zmax: 209.8951
#> Projected CRS: NAD83 / BC Albers
#> # A data frame: 20 × 3
#>    blue_line_key   rkm                      geometry
#>            <int> <dbl>                   <POINT [m]>
#>  1     356308001     0    Z (1290725 469483.5 9.626)
#>  2     356308001     1   Z (1291542 469173.9 10.131)
#>  3     356308001     2      Z (1292439 468922.9 7.5)
#>  4     356308001     3    Z (1292909 468665.1 9.687)
#>  5     356308001     4    Z (1293268 467977.4 9.784)
#>  6     356308001     5    Z (1293158 467231.2 9.784)
#>  7     356308001     6    Z (1293794 467668.3 9.784)
#>  8     356308001     7    Z (1294079 467938.4 9.784)
#>  9     356308001     8       Z (1294497 468239.5 10)
#> 10     356308001     9     Z (1294708 468081.9 10.5)
#> 11     356308001    10     Z (1295576 467912.1 10.5)
#> 12     356308001    11     Z (1296011 467111.8 10.5)
#> 13     356308001    12     Z (1296414 466358.5 10.5)
#> 14     356308001    13     Z (1297305 465976.3 10.5)
#> 15     356308001    14     Z (1298145 465435.2 10.5)
#> 16     356308001    15     Z (1298990 464903.5 10.5)
#> 17     356308001    16     Z (1299886 465109.3 11.5)
#> 18     356308001    17  Z (1300649 465508.7 48.0165)
#> 19     356308001    18 Z (1301136 464758.4 161.1287)
#> 20     356308001    19 Z (1301656 464161.2 209.8951)