Snap River Meter to Point

Assigns closest river meter to each spatial point. If the blue line key (blk) is missing then it is also assigned together with the distance to the river meter (distance_to_rm) in m.

Usage,

fwa_snap_rm_to_point(x, rm, ...)

Arguments

x

An sf object of spatial points with optional integer column blk.

rm

An sf object of spatial point with blk and rm columns.

...

Additional columns to group by when assigning. Elements with missing values in rm are assigned to any value in x.

Value

An updated version of x with integer columns blk and rm and numeric column distance_to_rm.

Examples

if (FALSE) {
rm <- fwa_rm(blue_line_key = 356308001, interval = 1000)
fwa_snap_rm_to_point(rm[2:4,], rm[3:5,])
}